+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalCheckoutRequest.kt
@@ -104,6 +104,7 @@ class PayPalCheckoutRequest @JvmOverloads constructor(
     override var recurringBillingPlanType: PayPalRecurringBillingPlanType? = null,
     var amountBreakdown: AmountBreakdown? = null,
     override var shouldOfferCredit: Boolean = false,
+    override var campaigns: List<PayPalCampaign> = emptyList(),
 ) : PayPalRequest(
     hasUserLocationConsent = hasUserLocationConsent,
     localeCode = localeCode,
@@ -117,7 +118,8 @@ class PayPalCheckoutRequest @JvmOverloads constructor(
     riskCorrelationId = riskCorrelationId,
     enablePayPalAppSwitch = enablePayPalAppSwitch,
     userAuthenticationEmail = userAuthenticationEmail,
-    lineItems = lineItems
+    lineItems = lineItems,
+    campaigns = campaigns
 ) {
 
     @OptIn(ExperimentalBetaApi::class)
@@ -192,6 +194,12 @@ class PayPalCheckoutRequest @JvmOverloads constructor(
             val jsonLineItems = JSONArray()
             lineItems.forEach { jsonLineItems.put(it.toJson()) }
             parameters.put(LINE_ITEMS_KEY, jsonLineItems)
+        }
+
+        if (campaigns.isNotEmpty()) {
+            val jsonCampaigns = JSONArray()
+            campaigns.forEach { jsonCampaigns.put(it.toJson()) }
+            parameters.put(PAYPAL_CAMPAIGNS_KEY, jsonCampaigns)
         }
 
         val experienceProfile = JSONObject()